fix(mocks): Extracted logger mock out to separate file so server doesn't try to import jasmine at runtime
