+++ src/server/bootstrap/services.bootstrapper.spec.ts
@@ -1,7 +1,7 @@
 import { addProviders } from '@angular/core/testing';
 import { Injectable } from '@angular/core';
 import { Logger } from '../../common/services/logger.service';
-import { LoggerMock } from '../../common/services/logger.service.spec';
+import { LoggerMock } from '../../common/services/logger.service.mock';
 import { bootstrap, BootstrapResponse } from './index';
 import { registry } from '../../common/registry/entityRegistry';
 import { AbstractService } from '../../common/services/service';
