+++ src/server/controllers/resource.controller.spec.ts
@@ -3,7 +3,7 @@ import { Injectable } from '@angular/core';
 import { ResourceController } from './resource.controller';
 import { Logger } from '../../common/services/logger.service';
 import { Server, RouteConfig } from '../servers/abstract.server';
-import { LoggerMock } from '../../common/services/logger.service.spec';
+import { LoggerMock } from '../../common/services/logger.service.mock';
 import { AbstractModel, identifier } from '../../common/models/model';
 import { ServerMock } from '../servers/abstract.server.spec';
 import { MockStore } from '../../common/stores/mock.store';
