+++ src/server/services/database.service.spec.ts
@@ -3,7 +3,7 @@ import { Database } from './database.service';
 import { addProviders, inject, async } from '@angular/core/testing';
 import * as typeorm from 'typeorm';
 import { RemoteCli } from './remoteCli.service';
-import { LoggerMock } from '../../common/services/logger.service.spec';
+import { LoggerMock } from '../../common/services/logger.service.mock';
 import { RemoteCliMock } from './remoteCli.service.mock';
 import { registry } from '../../common/registry/entityRegistry';
 import Spy = jasmine.Spy;
